backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn==0.27.1

# Database
SQLAlchemy==2.0.25
aiosqlite==0.19.0

# Weather API
python-dotenv==1.0.1
httpx==0.26.0

# Audio Processing (we already use these)
librosa==0.10.1
numpy==1.26.3
pandas==2.2.0

# Utilities
python-multipart==0.0.9
pydantic==2.6.1

# CORS
fastapi-cors==0.0.6